backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.4",
    "django-environ>=0.13.0",
    "django-rest-framework>=0.1.0",
    "django-simple-history>=3.11.0",
    "psycopg[binary]>=3.3.3",
]
